refactor to navigator observer
