+++ flutter/lib/src/sentry_flutter.dart
@@ -13,7 +13,6 @@ import 'integrations/connectivity/connectivity_integration.dart';
 import 'integrations/screenshot_integration.dart';
 import 'native/factory.dart';
 import 'native/native_scope_observer.dart';
-import 'navigation/navigation_timing_manager.dart';
 import 'profiling.dart';
 import 'renderer/renderer.dart';
 import 'native/sentry_native.dart';
@@ -233,13 +232,13 @@ mixin SentryFlutter {
   static void reportInitiallyDisplayed(BuildContext context) {
     final routeName = ModalRoute.of(context)?.settings.name;
     if (routeName != null) {
-      NavigationTimingManager().reportInitiallyDisplayed(routeName);
+      SentryNavigatorObserver().reportInitiallyDisplayed(routeName);
     }
   }
 
   /// Reports the time it took for the screen to be fully displayed.
   static void reportFullyDisplayed() {
-    NavigationTimingManager().reportFullyDisplayed();
+    SentryNavigatorObserver().reportFullyDisplayed();
   }
 
   @internal
